bug crushing
