+++ data-samples.js
@@ -1,5 +1,46 @@
+import { newUserObj } from "./data-structures.js";
+import { addObjectToUser } from "./data-functions.js";
+
+// Generate Sample Users
+export const sampleUser1 = () => {
+    return constructSampleUser(sampleUserData1);
+};
+export const sampleUser2 = () => {
+    return constructSampleUser(sampleUserData2);
+};
+
+// Sample User Constructor
+const constructSampleUser = (_sampleUserData) => {
+    const user = newUserObj(
+      _sampleUserData.name,
+      _sampleUserData.username,
+      _sampleUserData.description,
+      _sampleUserData.imgURL,
+      _sampleUserData.location,
+      _sampleUserData.email,
+      _sampleUserData.website,
+      _sampleUserData.twitter
+    );
+  
+    sampleOrganizations.forEach(org => {
+      addObjectToUser(org, user.orgs);
+    });
+    sampleRepos.forEach(repo => {
+      addObjectToUser(repo, user.repoData);
+    });
+    sampleProjects.forEach(proj => {
+      addObjectToUser(proj, user.projectsData);
+    });
+    samplePackages.forEach(pack => {
+      addObjectToUser(pack, user.packagesData);
+    });
+  
+    return user;
+  };
+
+
 // User Data
-export const sampleUserData1 = {
+const sampleUserData1 = {
     name: "Yue Zer",
     username: "yue_zer",
     description: "Software developer",
@@ -11,7 +52,7 @@ export const sampleUserData1 = {
     twitter: "@yue_zer"
 };
 
-export const sampleUserData2 = {
+const sampleUserData2 = {
     name: "John Doe",
     username: "doe_j",
     description: "Business executive",
@@ -26,7 +67,7 @@ export const sampleUserData2 = {
 
 
 // Organizations
-export const sampleOrganizations = [
+const sampleOrganizations = [
     {
         name: "Friendly Folks Org",
         img: "https://i0.wp.com/raulotaolea.com/wp-content/uploads/2018/02/header_blog_800x469-1.jpg?fit=730%2C428&ssl=1"
@@ -39,7 +80,7 @@ export const sampleOrganizations = [
 
 
 // Repos
-export const sampleRepos = [
+const sampleRepos = [
     {
         repoTitle: "Repo NUmber 1",
         description: "Test for repo 1",
@@ -63,7 +104,7 @@ export const sampleRepos = [
 ];
 
 // Projects
-export const sampleProjects = [
+const sampleProjects = [
     {
         title: "Project Title",
         description: "This is a description.",
@@ -75,7 +116,7 @@ export const sampleProjects = [
 ];
 
 // Packages
-export const samplePackages = [
+const samplePackages = [
     {
         titleImage: "",
         title: "",
